+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 # 1.4.0
 
 - Added an Activity area to each issue, including a Comments and a History tab (see usage guide for more details)
+- Fixed the card context menu occasionally closing itself right after opening on a card that needed scrolling into view
 
 # 1.3.3